+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/scope/DebugLocalScope.java
@@ -6,6 +6,7 @@ import com.oracle.truffle.api.frame.MaterializedFrame;
 import com.oracle.truffle.api.interop.InteropLibrary;
 import com.oracle.truffle.api.interop.UnknownIdentifierException;
 import com.oracle.truffle.api.interop.UnsupportedMessageException;
+import com.oracle.truffle.api.library.CachedLibrary;
 import com.oracle.truffle.api.library.ExportLibrary;
 import com.oracle.truffle.api.library.ExportMessage;
 import com.oracle.truffle.api.source.SourceSection;
@@ -77,11 +78,13 @@ public class DebugLocalScope implements EnsoObject {
             && this.bindingsByLevelsIdx < this.bindingsByLevels.size());
   }
 
+  @TruffleBoundary
   public static DebugLocalScope createFromFrame(EnsoRootNode rootNode, MaterializedFrame frame) {
     return new DebugLocalScope(
         rootNode, frame, gatherBindingsByLevels(rootNode.getLocalScope().flattenBindings()), 0);
   }
 
+  @TruffleBoundary
   private static DebugLocalScope createParent(DebugLocalScope childScope) {
     return new DebugLocalScope(
         childScope.rootNode,
@@ -137,6 +140,7 @@ public class DebugLocalScope implements EnsoObject {
 
   /** Returns the members from the current local scope and all the parent scopes. */
   @ExportMessage
+  @TruffleBoundary
   ScopeMembers getMembers(boolean includeInternal) {
     List<String> members = new ArrayList<>();
     bindingsByLevels.stream().skip(bindingsByLevelsIdx).forEach(members::addAll);
@@ -144,6 +148,7 @@ public class DebugLocalScope implements EnsoObject {
   }
 
   @ExportMessage
+  @TruffleBoundary
   boolean isMemberModifiable(String memberName) {
     return allBindings.containsKey(memberName);
   }
@@ -170,6 +175,7 @@ public class DebugLocalScope implements EnsoObject {
   }
 
   @ExportMessage
+  @TruffleBoundary
   boolean isMemberReadable(String memberName) {
     // When a value in a frame is null, it means that the corresponding
     // AssignmentNode was not run yet, and the slot kind of the
@@ -179,13 +185,15 @@ public class DebugLocalScope implements EnsoObject {
   }
 
   @ExportMessage
-  Object readMember(String member) {
+  @TruffleBoundary
+  Object readMember(String member, @CachedLibrary("this") InteropLibrary interop) {
     FramePointer framePtr = allBindings.get(member);
     var value = getValue(frame, framePtr);
     return value != null ? value : DataflowError.UNINITIALIZED;
   }
 
   @ExportMessage
+  @TruffleBoundary
   void writeMember(String member, Object value) throws UnknownIdentifierException {
     if (!allBindings.containsKey(member)) {
       throw UnknownIdentifierException.create(member);
@@ -225,6 +233,7 @@ public class DebugLocalScope implements EnsoObject {
   }
 
   @ExportMessage
+  @TruffleBoundary
   SourceSection getSourceLocation() {
     return rootNode.getSourceSection();
   }
@@ -236,6 +245,7 @@ public class DebugLocalScope implements EnsoObject {
   }
 
   @Override
+  @TruffleBoundary
   public String toString() {
     return String.format(
         "DebugLocalScope{rootNode = '%s', bindingsByLevels = %s, idx = %d}",